;ref: move DocFiles helpers to hledger-lib
