+++ hledger-lib/hledger-lib.cabal
@@ -106,6 +106,7 @@ library
       Hledger.Reports.PostingsReport
       Hledger.Utils
       Hledger.Utils.Debug
+      Hledger.Utils.DocFiles
       Hledger.Utils.IO
       Hledger.Utils.Parse
       Hledger.Utils.Regex
@@ -160,6 +161,7 @@ library
     , tasty >=1.2.3
     , tasty-hunit >=0.10.0.2
     , template-haskell
+    , temporary
     , terminal-size >=0.3.3
     , text >=1.2.4.1
     , time >=1.5
@@ -220,6 +222,7 @@ test-suite doctest
     , tasty >=1.2.3
     , tasty-hunit >=0.10.0.2
     , template-haskell
+    , temporary
     , terminal-size >=0.3.3
     , text >=1.2.4.1
     , time >=1.5
@@ -282,6 +285,7 @@ test-suite unittest
     , tasty >=1.2.3
     , tasty-hunit >=0.10.0.2
     , template-haskell
+    , temporary
     , terminal-size >=0.3.3
     , text >=1.2.4.1
     , time >=1.5
